requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
psycopg2-binary>=2.9.0
alembic==1.13.1

# Dependências de teste
pytest==8.3.3
pytest-cov==6.0.0
pytest-flask==1.3.0
pytest-mock==3.14.0
factory-boy==3.3.1

